backend/requirements.txt
Django>=6.0
djangorestframework>=3.16
djangorestframework_simplejwt>=5.5
django-cors-headers>=4.9
Pillow>=12.0
pypdf>=5.0
requests>=2.32
python-dotenv>=1.2
